chatting-organs-pipeline/pyproject.toml
[project]
name = "ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "crewai[tools]>=1.9.3",
    "elevenlabs>=2.34.0",
    "google-genai>=1.62.0",
    "python-dotenv>=1.1.1",
]
